refact: splitted Core and Platform primitives
