+++ src/SDK/Platform/Platform.d.ts
@@ -1,0 +1,3 @@
+export declare namespace Platform {
+
+}
