+++ src/SDK/Platform/Platform.ts
@@ -1,0 +1,7 @@
+// @ts-ignore
+import {default as _DashPlatformProtocol} from '@dashevo/dpp';
+
+export namespace Platform {
+    export let DashPlatformProtocol = _DashPlatformProtocol;
+}
+export {Platform as default};
