+++ src/SDK/Platform/index.ts
@@ -1,0 +1,1 @@
+export {Platform} from './Platform';
