+++ src/SDK/SDK.ts
@@ -1,7 +1,7 @@
 import {Client as _Client} from './Client';
+import {Core as _Core} from './Core';
+import {Platform as _Platform} from './Platform';
 import {default as _DAPIClient} from '@dashevo/dapi-client';
-// @ts-ignore
-import {default as _DashPlatformProtocol} from '@dashevo/dpp';
 
 import {
     Wallet as _Wallet,
@@ -13,27 +13,12 @@ import {
     plugins as _WalletLibPlugins
 } from '@dashevo/wallet-lib';
 
-import {
-    Transaction as _Transaction,
-    Address as _Address,
-    Block as _Block,
-    UnspentOutput as _UnspentOutput,
-    HDPublicKey as _HDPublicKey,
-    HDPrivateKey as _HDPrivateKey,
-    Mnemonic as _Mnemonic,
-    Network as _Network,
-    Input as _Input,
-    Output as _Output,
-    Script as _Script,
-    PublicKey as _PublicKey,
-    PrivateKey as _PrivateKey
-} from '@dashevo/dashcore-lib';
-
 export namespace SDK {
+    export let DAPIClient = _DAPIClient;
     export let Client = _Client;
 
-    export let DAPIClient = _DAPIClient;
-    export let DashPlatformProtocol = _DashPlatformProtocol;
+    export let Core = _Core;
+    export let Platform = _Platform;
 
     // Wallet-lib primitives
     export let Wallet = _Wallet;
@@ -47,21 +32,6 @@ export namespace SDK {
         utils: _WalletLibUtils,
     }
 
-    // Dashcore-lib primitives
-    export let Transaction = _Transaction;
-
-    export let Address = _Address;
-    export let Block = _Block;
-    export let UnspentOutput = _UnspentOutput;
-    export let HDPublicKey = _HDPublicKey;
-    export let HDPrivateKey = _HDPrivateKey;
-    export let PublicKey = _PublicKey;
-    export let PrivateKey = _PrivateKey;
-    export let Mnemonic = _Mnemonic;
-    export let Network = _Network;
-    export let Script = _Script;
-    export let Input = _Input;
-    export let Output = _Output;
 }
